Login, Sign Up, and Dashboard using Supabase
The default page for the website is the login page, which is also routed as /login. The user can signup in the signup page. All text fields have to be entered correctly and the user verifies their account, or if they forgot their password, through an email. All authentication is handled through Supabase. The user is presented with their email in the dashboard, and there is a button to sign out. The user is redirected from the dashboard page to the login page if they are not currently logged in.